+++ src/main/java/com/extia/cinema/services/ICategoryService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface ICategoryService {
 
     List<Category> findAll();
-    Category findById(Integer id);
+    Category findById(Long id);
     Category create(Category category);
     Category update(Category category);
-    void delete(Integer id);
+    void delete(Long id);
 }
